PC-ANOVA Result Methods

Standard result computation and extraction functions for ASCA (pcanova).

Usage

# S3 method for class 'pcanova'
summary(object, ...)

# S3 method for class 'summary.pcanova'
print(x, digits = 2, ...)

# S3 method for class 'pcanova'
print(x, ...)

# S3 method for class 'pcanova'
summary(object, ...)

Arguments

object

pcanova object.

...

additional arguments to underlying methods.

x

pcanova object.

digits

integer number of digits for printing.

Value

Returns depend on method used, e.g. projections.pcanova returns projected samples, scores.pcanova return scores, while print and summary methods return the object invisibly.

Details

Usage of the functions are shown using generics in the examples in pcanova. Explained variances are available (block-wise and global) through blockexpl and print.rosaexpl. Object printing and summary are available through: print.pcanova and summary.pcanova. Scores and loadings have their own extensions of scores() and loadings() through scores.pcanova and loadings.pcanova. Special to ASCA is that scores are on a factor level basis, while back-projected samples have their own function in projections.pcanova.

References

Luciano G, Næs T. Interpreting sensory data by combining principal component analysis and analysis of variance. Food Qual Prefer. 2009;20(3):167-175.

See also